Potential fix for code scanning alert no. 2: Implicit narrowing conversion in compound assignment

+++ src/main/java/com/replaymod/render/gui/GuiVideoRenderer.java
@@ -116,7 +116,7 @@ public class GuiVideoRenderer extends GuiScreen implements Tickable {
     }
 
     //the total render time
-    private int renderTimeTaken = 0;
+    private long renderTimeTaken = 0;
 
     //the time at which the Screen was last updated
     private long prevTime = -1;
